# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface WasteQuestionInput {
  /**
     * @minLength 3
     * @maxLength 500
     */
  question: string;
}

export interface WasteEvidence {
  wasteItem: string;
  category: string;
  disposalMethod: string;
  reason: string;
  safetyGuidance: string;
  sustainabilityImpact: string;
  relevanceScore: number;
}

export interface WasteAnalysis {
  question: string;
  category: string;
  recommendedDisposal: string;
  why: string;
  safetyGuidance: string;
  sustainabilityImpact: string;
  response: string;
  aiGenerated: boolean;
  aiProvider: string;
  retrievalCount: number;
  evidence: WasteEvidence[];
}

export interface WasteExamples {
  examples: string[];
}