# lib/api-zod/src/generated/types/wasteAnalysis.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WasteEvidence } from './wasteEvidence';

export interface WasteAnalysis {
  question: string;
  category: string;
  recommendedDisposal: string;
  why: string;
  safetyGuidance: string;
  sustainabilityImpact: string;
  response: string;
  aiGenerated: boolean;
  aiProvider: string;
  retrievalCount: number;
  evidence: WasteEvidence[];
}