# lib/api-zod/src/generated/types/wasteEvidence.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WasteEvidence {
  wasteItem: string;
  category: string;
  disposalMethod: string;
  reason: string;
  safetyGuidance: string;
  sustainabilityImpact: string;
  relevanceScore: number;
}