# lib/api-zod/src/generated/types/wasteQuestionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WasteQuestionInput {
  /**
     * @minLength 3
     * @maxLength 500
     */
  question: string;
}